Implemented first version of trainer TODO: derive patterns from corect answer rather than given ansver TODO: Store children as dictionary for faster lookup TODO: More elaborate selection and sorting of update-actions TODO: Train on increasingly more test cases until all are covered TODO: Do/Undo of update-actions TODO: Load and save generated test cases

+++ Grammar.Test/EvaluatorTests.cs
@@ -28,8 +28,8 @@ namespace Lingua.Grammar.Test
         [TestCase("^Vi", "^Vi")]
         public void MatchScorer(string pattern, string symbols)
         {
-            const int score = 1;
-            var evaluator = new Evaluator(new Dictionary<string, int> {{pattern, score } });
+            const sbyte score = 1;
+            var evaluator = new Evaluator(new Dictionary<string, sbyte> {{pattern, score } });
             var code = Encoder.Encode(symbols);
             var actualScore = evaluator.Evaluate(code).Score;
             Assert.That(actualScore, Is.EqualTo(score));
